Move types and definitions to sub folder
Better organize the code.
