+++ src/definitions/FrameIdentifiers.ts
@@ -211,106 +211,3 @@ export const FRAME_ALIASES =
         },
         {}
     )
-
-export type FrameOptions = {
-    multiple: boolean
-    updateCompareKey?: string
-}
-
-const satisfiesFrameOptions =
-    <T extends Record<string, FrameOptions>>(data: T) => data
-
-export const ID3_FRAME_OPTIONS = satisfiesFrameOptions({
-    "PIC": {
-        multiple: false /* change in 1.0 */
-    },
-    "WCM": {
-        multiple: true
-    },
-    "WAR": {
-        multiple: true
-    },
-    "T___": {
-        // This is "correct", but in v4, the text frame's value can be split by using 0x00.
-        // https://github.com/Zazama/node-id3/issues/111
-        multiple: false
-    },
-    "TXXX": {
-        multiple: true,
-        updateCompareKey: "description"
-    },
-    "APIC": {
-        multiple: false /* change in 1.0 */
-    },
-    "USLT": {
-        multiple: false /* change in 1.0 */
-    },
-    "SYLT": {
-        multiple: true
-    },
-    "COMM": {
-        multiple: false /* change in 1.0 */
-    },
-    "POPM": {
-        multiple: false /* change in 1.0 */
-    },
-    "PRIV": {
-        multiple: true
-    },
-    "CTOC": {
-        multiple: true
-    },
-    "CHAP": {
-        multiple: true
-    },
-    "WXXX": {
-        multiple: true,
-        updateCompareKey: "description"
-    },
-    "WCOM": {
-        multiple: true
-    },
-    "WOAR": {
-        multiple: true
-    },
-    "ETCO": {
-        multiple: false
-    },
-    "COMR": {
-        multiple: true
-    },
-    "UFID": {
-        multiple: true
-    }
- } as const)
-
-/*
-**  Officially available types of the picture frame
-*/
-export const APIC_TYPES = [
-    "other",
-    "file icon",
-    "other file icon",
-    "front cover",
-    "back cover",
-    "leaflet page",
-    "media",
-    "lead artist",
-    "artist",
-    "conductor",
-    "band",
-    "composer",
-    "lyricist",
-    "recording location",
-    "during recording",
-    "during performance",
-    "video screen capture",
-    "a bright coloured fish",
-    "illustration",
-    "band logotype",
-    "publisher logotype"
-] as const
-
-export const ENCODINGS = [
-    'ISO-8859-1', 'UTF-16', 'UTF-16BE', 'utf8'
-] as const
