+++ src/definitions/PictureTypes.ts
@@ -1,0 +1,26 @@
+/**
+ *  Officially available types of the picture frame
+ */
+ export const APIC_TYPES = [
+    "other",
+    "file icon",
+    "other file icon",
+    "front cover",
+    "back cover",
+    "leaflet page",
+    "media",
+    "lead artist",
+    "artist",
+    "conductor",
+    "band",
+    "composer",
+    "lyricist",
+    "recording location",
+    "during recording",
+    "during performance",
+    "video screen capture",
+    "a bright coloured fish",
+    "illustration",
+    "band logotype",
+    "publisher logotype"
+] as const
